+++ bluetooth_low_energy_android/lib/src/my_peripheral_manager.dart
@@ -156,9 +156,28 @@ final class MyPeripheralManager extends PlatformPeripheralManager
 
   @override
   Future<void> startAdvertising(Advertisement advertisement) async {
-    final advertisementArgs = advertisement.toArgs();
-    logger.info('startAdvertising: $advertisementArgs');
-    await _api.startAdvertising(advertisementArgs);
+    final nameArgs = advertisement.name;
+    if (nameArgs != null) {
+      logger.info('setName: $nameArgs');
+      final newNameArgs = await _api.setName(nameArgs);
+      if (newNameArgs != nameArgs) {
+        throw ArgumentError(
+            'Name changed, but $newNameArgs is different from $nameArgs');
+      }
+    }
+    final settingsArgs = MyAdvertiseSettingsArgs(
+      modeArgs: MyAdvertiseModeArgs.balanced,
+      connectableArgs: true,
+    );
+    final advertiseDataArgs = advertisement.toAdvertiseDataArgs();
+    final scanResponseArgs = advertisement.toScanResponseArgs();
+    logger.info(
+        'startAdvertising: $settingsArgs, $advertiseDataArgs, $scanResponseArgs');
+    await _api.startAdvertising(
+      settingsArgs,
+      advertiseDataArgs,
+      scanResponseArgs,
+    );
   }
 
   @override