+++ bluetooth_low_energy/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 6.0.1
+
+* `Android` Fix the issue that [advertisement name is wrong when advertising](https://github.com/yanshouwang/bluetooth_low_energy/issues/62).
+
 ## 6.0.0
 
 * [Add `CentralManager#retrieveConnectedPeripherals` method.](https://github.com/yanshouwang/bluetooth_low_energy/issues/61)