+++ bluetooth_low_energy_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 6.0.2
+
+* Fix the warning issue.
+
+## 6.0.1
+
+* Fix the issue that [advertisement name is wrong when advertising](https://github.com/yanshouwang/bluetooth_low_energy/issues/62).
+
 ## 6.0.0
 
 * Add `serviceUUIDs` argument to `CentralManager#startDiscovery` method.